+++ src/dom-renderer.js
@@ -95,7 +95,7 @@ export default function DOMRenderer(game, boardElement) {
       utils.appendElement(boardElement.querySelector(".lines.horizontal"), horizontalLine);
 
       const verticalLine = utils.createElement("div", { class: "line vertical" });
-      verticalLine.setAttribute("data-top-gutter", game.xCoordinateFor(y))
+      verticalLine.setAttribute("data-top-gutter", game.xCoordinateFor(y));
       utils.appendElement(boardElement.querySelector(".lines.vertical"), verticalLine);
 
       for (let x = 0; x < game.boardSize; x++) {
@@ -148,12 +148,20 @@ export default function DOMRenderer(game, boardElement) {
         const intersectionElement = this;
 
         utils.addClass(intersectionElement, "hovered");
+
+        const hoveredYPosition = Number(intersectionElement.getAttribute("data-position-y"));
+        const hoveredXPosition = Number(intersectionElement.getAttribute("data-position-x"));
+
+        if (game.isIllegalAt(hoveredYPosition, hoveredXPosition)) {
+          utils.addClass(intersectionElement, "illegal");
+        }
       });
 
       utils.addEventListener(intersectionEl, "mouseleave", function() {
         const intersectionElement = this;
 
         utils.removeClass(intersectionElement, "hovered");
+        utils.removeClass(intersectionElement, "illegal");
         renderer.resetTouchedPoint();
       });
 
@@ -285,7 +293,7 @@ export default function DOMRenderer(game, boardElement) {
       renderer.lastTranslateX = translateX;
       renderer.lastTranslateY = translateY;
     });
-  }
+  };
 
   this.showPossibleMoveAt = function(intersectionElement) {
     const renderer = this;
@@ -314,7 +322,7 @@ export default function DOMRenderer(game, boardElement) {
     const renderer = this;
 
     renderer.touchedPoint = null;
-  }
+  };
 
   this.zoomOut = function() {
     const renderer = this;
@@ -344,7 +352,7 @@ export default function DOMRenderer(game, boardElement) {
     if (this.game.isOver()) {
       this.renderTerritory();
     }
-  }
+  };
 
   this.renderStonesPlayed = function() {
     const renderer = this;
@@ -362,12 +370,6 @@ export default function DOMRenderer(game, boardElement) {
     if (!boardState) {
       return;
     }
-
-    renderer.game.intersections().forEach(function(intersection) {
-      if (renderer.game.wouldBeSuicide(intersection.y, intersection.x)) {
-        utils.addClass(renderer.grid[intersection.y][intersection.x], "suicide");
-      }
-    });
 
     if (boardState.koPoint) {
       utils.addClass(renderer.grid[boardState.koPoint.y][boardState.koPoint.x], "ko");