+++ src/game.js
@@ -1,7 +1,14 @@
 import DOMRenderer from "./dom-renderer";
 import NullRenderer from "./null-renderer";
 import BoardState from "./board-state";
-import { TerritoryRules, AreaRules } from "./rulesets";
+import Ruleset from "./ruleset";
+
+const VALID_GAME_OPTIONS = [
+  "boardSize",
+  "scoring",
+  "handicapStones",
+  "koRule"
+];
 
 const Game = function(boardElement) {
   this._defaultBoardSize = 19;
@@ -12,11 +19,12 @@ const Game = function(boardElement) {
     postRender: function() {}
   };
   this._deadPoints = [];
-  this._defaultRuleset = "territory";
+  this._defaultScoring = "territory";
+  this._defaultKoRule = "simple";
 };
 
 Game.prototype = {
-  _configureOptions: function({ boardSize = this._defaultBoardSize, handicapStones = 0, ruleset = this._defaultRuleset } = {}) {
+  _configureOptions: function({ boardSize = this._defaultBoardSize, handicapStones = 0, scoring = this._defaultScoring, koRule = this._defaultKoRule } = {}) {
     if (handicapStones > 0 && boardSize !== 9 && boardSize !== 13 && boardSize !== 19) {
       throw new Error("Handicap stones not supported on sizes other than 9x9, 13x13 and 19x19");
     }
@@ -27,17 +35,19 @@ Game.prototype = {
 
     this.boardSize = boardSize;
     this.handicapStones = handicapStones;
-    this.ruleset = {
-      "area": AreaRules,
-      "territory": TerritoryRules
-    }[ruleset];
-
-    if (!this.ruleset) {
-      throw new Error("Unknown ruleset: " + ruleset);
-    }
+    this.ruleset = new Ruleset({
+      "scoring": scoring,
+      "koRule": koRule
+    });
   },
 
   setup: function(options) {
+    for (let key in options) {
+      if (options.hasOwnProperty(key) && VALID_GAME_OPTIONS.indexOf(key) < 0) {
+        throw new Error("Unrecognized game option: " + key);
+      }
+    }
+
     this._configureOptions(options);
 
     if (this.boardSize > 19) {
@@ -107,7 +117,7 @@ Game.prototype = {
 
   pass: function() {
     if (!this.isOver()) {
-      this._moves.push(this.boardState().playPass())
+      this._moves.push(this.boardState().playPass());
       this.render();
     }
   },
@@ -162,7 +172,7 @@ Game.prototype = {
       return false;
     }
 
-    return this.ruleset.isIllegal(y, x, this.boardState());
+    return this.ruleset.isIllegal(y, x, this);
   },
 
   render: function() {